:recycle: refactor: Ajustes telas de notificações
