+++ package.json
@@ -48,7 +48,7 @@
     "react": "^18.2.0",
     "react-chartjs-2": "^5.2.0",
     "react-dom": "^18.2.0",
-    "react-icons": "^4.7.1",
+    "react-icons": "^4.12.0",
     "react-input-mask": "^3.0.0-alpha.2",
     "react-leaflet": "^4.2.1",
     "react-qr-code": "^2.0.11",
